refactor: move grug/vm/wasm tests to dango/testing
Moves tester.rs and transfer.rs into dango/testing/tests/vm/wasm.rs.
Copies wasm testdata files to dango/testing/testdata/. Removes
grug-testing from grug-vm-wasm dev-deps (keeps deps needed for
inline tests and benchmarks).

---------
