+++ dango/testing/tests/vm/main.rs
@@ -1,0 +1,1 @@
+mod wasm;
